fix(ui): harden docking refactor gates
Repair scroll/text semantics invalidation, active-descendant timing, and gallery overlay regression gates after the docking refactor.

Keep shadcn/material3 snapshot surfaces and advanced public API checks aligned with the updated contracts.

+++ crates/fret-ui/src/tree/tests/window_command_action_availability_snapshot.rs
@@ -1490,7 +1490,7 @@ fn action_availability_no_focus_subtree_fallback_reuses_subtree_interest_across_
 
     assert_eq!(
         crate::tree::take_command_availability_subtree_interest_probe_count(),
-        (depth + 1) as usize,
+        depth + 1,
         "subtree interest should be computed once per node in the fallback subtree"
     );
